# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Service Deprecation & Migration Analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyticsSummary {
  totalUsers: number;
  activeUsers30d: number;
  deprecatedVersionUsers: number;
  migratedUsers: number;
  migrationRate: number;
  avgErrorRate: number;
  highChurnUsers: number;
  pendingMigrations: number;
}

export interface VersionCount {
  version: string;
  count: number;
  percentage: number;
  isDeprecated: boolean;
}

export interface FunnelStage {
  stage: string;
  count: number;
  percentage: number;
}

export interface RegionStats {
  region: string;
  totalUsers: number;
  activeUsers: number;
  avgErrorRate: number;
  v1Users: number;
  v2Users: number;
  v3Users: number;
}

export interface DailyActiveUsers {
  date: string;
  total: number;
  v1: number;
  v2: number;
  v3: number;
}

export interface ErrorTrendPoint {
  date: string;
  v1ErrorRate: number;
  v2ErrorRate: number;
  v3ErrorRate: number;
}

export interface ChurnRiskSegment {
  churnRisk: string;
  tier: string;
  count: number;
  avgErrorRate: number;
  avgDaysSinceActive: number;
}

export interface UserRecord {
  id: number;
  userId: string;
  serviceVersion: string;
  lastActiveDate: string;
  usageFrequency: string;
  region: string;
  errorRate: number;
  subscriptionTier: string;
  migrationStatus: string;
  churnRisk: string;
  daysSinceActive: number;
}

export interface UserListResponse {
  users: UserRecord[];
  total: number;
  page: number;
  limit: number;
}

export interface FrequencySegment {
  version: string;
  frequency: string;
  count: number;
}

export interface EventLogEntry {
  id: number;
  timestamp: string;
  userId: string;
  serviceVersion: string;
  region: string;
  eventType: string;
  /** @nullable */
  errorCode: string | null;
  latencyMs: number;
}

export interface ErrorSpike {
  timestamp: string;
  version: string;
  errorRate: number;
  affectedUsers: number;
  region: string;
}

export type GetDailyActiveUsersParams = {
days?: number;
};

export type GetErrorTrendsParams = {
days?: number;
};

export type GetUsersParams = {
version?: string;
region?: string;
tier?: string;
churnRisk?: string;
activeOnly?: boolean;
deprecated?: boolean;
page?: number;
limit?: number;
};

export type GetEventLogParams = {
days?: number;
limit?: number;
};