# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cloud Service Deprecation & Migration Analytics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telemetry/event-log:
    get:
      operationId: getEventLog
      tags: [telemetry]
      summary: Recent activity event log
      description: Simulated KQL-style event log for the last N days with version, region, event type
      parameters:
        - name: days
          in: query
          required: false
          schema:
            type: integer
            default: 7
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Event log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventLogEntry"
...
components:
  schemas:
...
    EventLogEntry:
      type: object
      required: [id, timestamp, userId, serviceVersion, region, eventType, errorCode, latencyMs]
      properties:
        id:
          type: integer
        timestamp:
          type: string
        userId:
          type: string
        serviceVersion:
          type: string
        region:
          type: string
        eventType:
          type: string
        errorCode:
          type: ["string", "null"]
        latencyMs:
          type: integer
...