# lib/api-zod/src/generated/types/eventLogEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud Service Deprecation & Migration Analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface EventLogEntry {
  id: number;
  timestamp: string;
  userId: string;
  serviceVersion: string;
  region: string;
  eventType: string;
  /** @nullable */
  errorCode: string | null;
  latencyMs: number;
}